Reimplement camera switching
